refactor: split sidecar proxy handlers per backend
